requirements.txt
pandas
numpy
notebook
matplotlib
nltk
tensorflow
pillow
huggingface-hub
datasets
wandb
dvc
ipykernel
python-box==6.0.2
tqdm
pyYAML
ensure==1.0.2
joblib
types-pyYAML
scipy
Flask
Flask-Cors
python-dotenv
kaggle
-e .